Add exclusive or into grouped scenario
