+++ editor.planx.uk/src/@planx/components/Checklist/Public/components/AutoAnsweredChecklist.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 
-import { Props } from "../../types";
+import { PublicChecklistProps } from "../../types";
 
 // An auto-answered Checklist won't be seen by the user, but still leaves a breadcrumb
 export const AutoAnsweredChecklist: React.FC<
-  Props & { answerIds: string[] }
+  PublicChecklistProps & { answerIds: string[] }
 > = (props) => {
   useEffect(() => {
     props.handleSubmit?.({
